LPD-57225 Move settings component to specific folder and create new file for entry point
